Move wasm annotations to wasm.minimal.attanotion
