+++ test-bridge/src/main/scala/org/scalajs/testing/bridge/JSRPC.scala
@@ -16,7 +16,7 @@ import scala.scalajs.js
 import scala.scalajs.js.annotation._
 import scala.scalajs.LinkingInfo._
 import scala.scalajs.LinkingInfo.ModuleKind.MinimalWasmModule
-import scala.scalajs.wasm.annotation._
+import scala.scalajs.wasm.minimal.annotation._
 
 import scala.collection.mutable
 import scala.concurrent.duration._
